Create AfterHoursVmCreation.ps1
This is intended to be ran as a scheduled task after hours. The goal is to create 110% of the max users totals per pool in an Instant-clone Vmware Horizon environment after hours. This will free up each host CPU from deleting and recreating clones during the day and allow for better cpu ready times for each vm users are using.